refactor: extract the shared daemon merge path from the switch commands
write and delete on an active profile need the same sequence — drift
gate, daemon merge with one interrupted retry, reload-and-replay
persistence, change notification — so it moves out of SwitchCommand
into MergeCommitFlow unchanged, with the switch-specific wording of two
error messages generalized.
